backend/requirements_final1.txt
# Core Framework
fastapi==0.121.0
uvicorn[standard]==0.38.0
gunicorn==23.0.0
python-multipart==0.0.20
python-dotenv==1.2.1
starlette==0.49.3

# Database
sqlalchemy==2.0.44

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib==1.7.4
cryptography==46.0.3
email-validator==2.3.0
pydantic==2.12.3
pydantic_core==2.41.4

# HTTP & WebSockets
httptools==0.7.1
websockets==15.0.1
requests==2.32.5
h11==0.16.0

# ML (ONLY scikit-learn for TF-IDF)
scikit-learn==1.7.2
numpy==2.3.4

# Document Processing
pdfplumber==0.11.7
pdfminer.six==20250506
pypdfium2==5.0.0

# Google OAuth
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.2.1

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
click==8.3.0
colorama==0.4.6
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0
typing_extensions==4.15.0
packaging==25.0
filelock==3.20.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.3
watchfiles==1.1.1